requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-groq
faiss-cpu
pymupdf
